requirements.txt
py-cord
requests
cryptography
pyyaml
tqdm
pygments